Address PR review feedback
- Derive the ArcadeLogging.dll toolset subfolder ('net' vs 'netframework')
  from the build tool command instead of \\.Framework\, which no
  longer exists after the MSBuild-function merge (#16916) and evaluated to
  \, breaking Join-Path when pipelinesLog was enabled.
- Check the logger assembly actually exists (Test-Path / -f) instead of the
  always-true non-empty-string check, so a missing logger fails fast with an
  actionable message.
- Inline the message formatting into PipelinesLogger and delete MessageBuilder.cs
  (per review): the state-machine builder was overkill for the minimal logger
  and carried unused usings that would trip TreatWarningsAsErrors.
- Fix comment typos in tools.ps1/tools.sh ('opt-opt' -> 'opt-in',
  'while validate' -> 'while validating').
